+++ netket/hilbert/random/spin.py
@@ -14,82 +14,51 @@
 
 import jax
 import numpy as np
-from jax import numpy as jnp
+from functools import partial
 
 from netket.hilbert import Spin
 from netket.utils.dispatch import dispatch
 
+from .fock import _random_states_with_constraint_fock
+
 
 @dispatch
 def random_state(hilb: Spin, key, batches: int, *, dtype=np.float32):
-    S = hilb._s
-    shape = (batches, hilb.size)
-
-    # If unconstrained space, use fast sampling
+    shape = (batches,)
     if hilb._total_sz is None:
-        n_states = int(2 * S + 1)
-        rs = jax.random.randint(key, shape=shape, minval=0, maxval=n_states)
-
-        two = jnp.asarray(2, dtype=dtype)
-        return jnp.asarray(rs * two - (n_states - 1), dtype=dtype)
+        return _random_states_unconstrained(hilb, key, shape, dtype)
     else:
-        N = hilb.size
-        n_states = int(2 * S) + 1
-        # if constrained and S == 1/2, use a trick to sample quickly
-        if n_states == 2:
-            m = int(hilb._total_sz * 2)
-            nup = (N + m) // 2
-            ndown = (N - m) // 2
-
-            x = jnp.concatenate(
-                (
-                    jnp.ones((batches, nup), dtype=dtype),
-                    -jnp.ones(
-                        (
-                            batches,
-                            ndown,
-                        ),
-                        dtype=dtype,
-                    ),
-                ),
-                axis=1,
-            )
-
-            return jax.vmap(jax.random.permutation)(
-                jax.random.split(key, x.shape[0]), x
-            )
-
-        # if constrained and S != 1/2, then use a slow fallback algorithm
-        # TODO: find better, faster way to sample constrained arbitrary spaces.
-        else:
-            state = jax.pure_callback(
-                lambda rng: _random_states_with_constraint(hilb, rng, batches, dtype),
-                jax.ShapeDtypeStruct(shape, dtype),
-                key,
-            )
-
-            return state
+        return _random_states_with_constraint(hilb, key, shape, dtype)
 
 
-# TODO: could numba-jit this
-def _random_states_with_constraint(hilb, rngkey, n_batches, dtype):
-    out = np.full((n_batches, hilb.size), -round(2 * hilb._s), dtype=dtype)
-    rgen = np.random.default_rng(np.asarray(rngkey))
+# For the implementations of constrained spaces, we use those found inside of
+# fock.py, and simply convert the spin values (-1, 1) to fock values (0,1, ...).
 
-    for b in range(n_batches):
-        sites = list(range(hilb.size))
-        ss = hilb.size
+_spin_to_fock = lambda two_times_s, x: (two_times_s + x) // 2
+_fock_to_spin = lambda two_times_s, x: 2 * x - two_times_s
 
-        for i in range(round(hilb._s * hilb.size + hilb._total_sz)):
-            s = rgen.integers(0, ss, size=())
 
-            out[b, sites[s]] += 2
+@partial(jax.jit, static_argnames=("hilb", "shape", "dtype"))
+def _random_states_unconstrained(hilb, key, shape, dtype):
+    two_times_s = round(2 * hilb._s)
+    x_fock = jax.random.randint(
+        key, shape=shape + (hilb.size,), minval=0, maxval=two_times_s + 1
+    )
+    return _fock_to_spin(two_times_s, x_fock).astype(dtype)
 
-            if out[b, sites[s]] > round(2 * hilb._s - 1):
-                sites.pop(s)
-                ss -= 1
 
-    return out
+@partial(jax.jit, static_argnames=("hilb", "shape", "dtype"))
+def _random_states_with_constraint(hilb, key, shape, dtype):
+    # Generate random spin states with a given hilb._total_sz.
+    # Note that this is NOT a uniform distribution over the
+    # basis states of the constrained hilbert space.
+    two_times_s = round(2 * hilb._s)
+    two_times_total_sz = round(2 * hilb._total_sz)
+    n_particles = _spin_to_fock(two_times_s * hilb.size, two_times_total_sz)
+    x_fock = _random_states_with_constraint_fock(
+        n_particles, hilb.shape, key, shape, dtype
+    )
+    return _fock_to_spin(two_times_s, x_fock).astype(dtype)
 
 
 @dispatch